change uint32 to uint64

+++ konigsberg/sparseutil.py
@@ -20,7 +20,7 @@ from numba.types import List, Tuple
 @nb.njit(nogil=True)
 def make_counts(maps, ptr_arr):
     """Count number of edges per index (in the domain) with accumulation.
-    
+
     In other words, for every input index, ptr_arr[index] is first set
     to the number of elements that index maps to. A cumulative sum is
     then performed over ptr_arr.
@@ -41,7 +41,7 @@ def make_counts(maps, ptr_arr):
     # Account for multiple maps. ptr_arr_2d[domain_index, map_index].
     ptr_arr_2d = ptr_arr[:-1].reshape(-1, number_maps)
     for i, map_ in enumerate(maps):
-        map_ptr = ptr_arr_2d[...,i]
+        map_ptr = ptr_arr_2d[..., i]
         for from_arr, _ in map_:
             # As if all arrays in map_ were concatenated.
             for j in from_arr:
@@ -62,7 +62,7 @@ def place_indices(maps, ptr_arr, ind_arr):
     number_maps = len(maps)
     ptr_arr_2d = ptr_arr[:-1].reshape(-1, number_maps)
     for i, map_ in enumerate(maps):
-        map_ptr = ptr_arr_2d[...,i]
+        map_ptr = ptr_arr_2d[..., i]
         for from_arr, to_arr in map_:
             # Treat all arrays in map_ as concatenated.
             for j, k in zip(from_arr, to_arr):
@@ -107,19 +107,19 @@ def make_sparse_matrix(n_from_ind, maps, ptr_path, ind_path):
     with open(ptr_path, 'wb+') as f_ptr:
         # Trick: seek past the end of file and write one byte to
         # efficiently zero-fill up to that point.
-        f_ptr.seek((n_maps * n_from_ind + 1) * np.uint32().nbytes - 1)
+        f_ptr.seek((n_maps * n_from_ind + 1) * np.uint64().nbytes - 1)
         f_ptr.write(b'\x00')
         f_ptr.flush()
         map_ptr = mmap.mmap(f_ptr.fileno(), 0)
     with map_ptr:
-        arr_ptr = np.frombuffer(map_ptr, dtype=np.uint32)
+        arr_ptr = np.frombuffer(map_ptr, dtype=np.uint64)
         make_counts(maps, arr_ptr)
         with open(ind_path, 'wb+') as f_ind:
             # Same trick.
-            f_ind.seek(n_links * np.uint32().nbytes - 1)
+            f_ind.seek(n_links * np.uint64().nbytes - 1)
             f_ind.write(b'\x00')
             f_ind.flush()
             map_ind = mmap.mmap(f_ind.fileno(), 0)
         with map_ind:
-            arr_ind = np.frombuffer(map_ind, dtype=np.uint32)
+            arr_ind = np.frombuffer(map_ind, dtype=np.uint64)
             place_indices(maps, arr_ptr, arr_ind)
